Add Custom Parameter Type Support
In addition, teach Burpless how to deal with `:keyword`s in feature
files and `{keyword}` parameter types in `CucumberExpression` strings.

Also, make a change to Burpless internals where, instead of maintaining
an internal map of parameter type string to java type, derive it instead
from the parameter type registry's internal map. This change is
necessary in order to have up-to-date `ParameterInfo` computations as
well.

Resolves #2

Signed-off-by: Daniel Miladinov <daniel.miladinov@gmail.com>