Saving this Snippet For a Rainy Day
Do we need to support parameterized DocString types?
Coming from Clojure, will we ever _really need_ to distinguish between,
say, for example, `Map<String, String>` from `Map<String, Object>`, or
`Map<Keyword, IObj>`, when the underlying raw type for all of these is
just `Map`?

This would only really be an issue if / when we need to support multiple
variants of parameterized maps, each with their own transform function,
but all having the same content-type, and saying that their :to-type is
just `Map` isn't specific enough…

Signed-off-by: Daniel Miladinov <daniel.miladinov@gmail.com>