Use Feature Pickle To Infer Step Arguments
Until now, writers of step fns would have to manually add a bit of
metadata to their step fns, indicating that the step took a
`^:datatable` or ` ^:docstring` argument.

That was a terrible developer experience, and I felt like there had to
be a better way. I could see that under the covers, the Cucumber runtime
would generate a `GherkinMessagesFeature` instance from the feature file
provided to it, and that `Feature` would contain all of the `Scenario`s,
each with all of their `Step`s, each with all of their `Argument`s,
either `DataTable` or `DocString`, for those that had any.

But at the point where I need to match up the Clojure step fns to the
feature and provide appropriate `StepDefinition` instances for each, I
lacked the necessary information to match up a specific fn to a specific
step. I didn't see any way from within the reified `StepDefinition` to
get at the caller to somehow get at the `Feature` pickle so as to
identify the step argument information.

So, in defeat, and as a hack, I punted this responsibility of
identifying the step fns that take an argument, either a `DataTable` or
`DocString`, to the writer of the step fn. I found that I could look up
more information about the step fn from its metadata map, if it had any.
And using that information from the metadata I could then append the
corresponding `DataTable` or `DocString` `ParameterInfo` to whatever
other `ParameterInfo` I extracted out of the step text pattern itself.

I had somehow convinced myself that this was the way things had to be.
And when I got some user feedback about how terrible this was, that
there needed to be a better experience around Clojure step fn snippets,
that didn't need any additional modification before they would be
correctly matched up to their step information from the feature, and
would be called at the correct time and with the correct arguments, I
felt almost certain there was nothing I could do without first changing
something within
[cucumber-jvm](https://github.com/cucumber/cucumber-jvm) itself.

I even went so far as to propose a change upstream to `cucumber-jvm`'s
`io.cucumber.core.backend.Snippet` interface, so that its `template`
method accepted the current step for which the template was being
generated. Then I could query the step for its argument, and
if present, I could inspect the type of the argument and determine
whether it was for a `DataTable` or `DocString`. Then I could a
`MessageFormat` generated from one of three different strings:
- one where the step fn bore no metadata
- one where the step fn bore a `^:datatable` metadata
- one where the step fn bore a `^:docstring` metadata

I thought this would be it, as it would achieve my goal of `burpless`
users no longer needing to manually edit their step fn snippets before
they would be picked up by the Cucumber runtime.

Alas, that change went nowhere as it was indeed a breaking change to a
public API and the maintainers didn't see any reason needing to accept
it. Their abrupt but totally reasonable and fair response forced me to
go back to the drawing board and see if there wasn't some other way to
achieve my goal.

It was then that I noticed, that if you build a Cucumber runtime without
specifying the `Feature` to use, the runtime would go ahead and create
one itself. But, you could also specify a `Feature` object for the
runtime to use. The `Feature` instance represents a parsed feature file,
and that `Feature` would contain all of the `Scenario`s,
each with all of their `Step`s, each with all of their `Argument`s,
either `DataTable` or `DocString`, for those that had any. If I have the
`Feature` data in hand, I could match up all the steps from the feature
with the step fns provided by the user! But how could I produce such a
`Feature` instance?

I studied the cucumber-jvm source in order to discover how it parses
feature files, and I discovered that:
1. It's a public API, and
2. I have all the data I need to produce one myself!

If I already knew the path to the feature file (which I do), then I
could generate a `java.net URI` from the incoming feature path `String`,
and from that `URI`, I could also derive a `java.nio.file.Path` from it,
with which I can then turn around and open an `InputStream` to the
feature file. With the `URI` and `InputStream` in hand, I could then new
up a parser and parse the contents of the `URI` / `InputStream`.

I believe the `InputStream` is what produces the contents of the
feature file, and the `URI` is only used for error reporting (“In which
file for which feature did this step(s) fail?”)

With the `Feature` instance in hand, I can then collect all of the
`Step`s from all of the `Scenario`s and use them to compare against
incoming step fns. I match them up by evaluating the `StepExpression`
for a given step fn against each step, where the matching step returns a
non-nil value.

I can then use the step argument to add the `ParameterInfo` for either
the `DataTable` or `DocString` as needed, without having to rely on the
step fn's metadata!

In order to eliminate duplication of effort, I take the `Feature`
instance I already parsed, and supply it to the runtime, so that it can
skip the feature parsing step itself.

Best of all, `burpless` remains backwards-compatible with existing step
fns that bear a `^:datatable` or `^:docstring` metadata!

And, for the most part, unless you're doing something specific with
either `DocStringType` or `DataTableType` transformations, you won't
ever need to supply those `^:datatable` or `^:docstring` metadata going
forward.

Resolves #25

Signed-off-by: Daniel Miladinov <daniel.miladinov@gmail.com>