Put `DocStringTypeDefinition`s To Good Use
We can use what we know about all the `DocStringTypeDefinition`s
currently loaded to deduce the final transformed type of a `DocString`'s
content, and fill in that type as the type hint when emitting step fn
snippets.

In addition, step fn snippets will no longer be emitted with a name, as
it's perfectly fine to leave the step fns anonymous. Personally I always
delete the fn name after copying from the snippet, anyway.  There was no
real need to include the fn name, as the pattern is already there, and
that is enough to sufficiently identify for which step the fn is.

Also, I fixed the minor whitespace issues around the arglist, making
sure there won't be trailing whitespace when there no other arguments
coming neither from the step nor any docstring or datatable.

Finally, we can also use `DocStringTypeDefinition`s to apply their
transformations to the appropriate `DocString` content (`String`) values
right in the step definitions, so that step fn authors can receive their
`DocString` content already transformed.

Signed-off-by: Daniel Miladinov <daniel.miladinov@gmail.com>